feat: ditch the api contract
