+++ packages/frontend2/src/app/(legacy)/zk-catalog/_utils/getProofVerification.ts
@@ -1,14 +1,11 @@
 import { type OnchainVerifier, type ProofVerification } from '@l2beat/config'
-import {
-  assert,
-  UnixTime,
-  type VerifiersApiResponse,
-} from '@l2beat/shared-pure'
+import { assert, UnixTime } from '@l2beat/shared-pure'
 import { type Project, type ZkCatalogProofVerification } from './types'
+import { type VerifierStatuses } from '~/server/features/zk-catalog/get-verifiers'
 
 export function getProofVerification(
   project: Project,
-  verifiersApiResponse: VerifiersApiResponse,
+  verifiersStatuses: VerifierStatuses,
 ): ZkCatalogProofVerification {
   let proofVerification: ProofVerification
 
@@ -24,19 +21,16 @@ export function getProofVerification(
     shortDescription: proofVerification.shortDescription,
     verifiers: proofVerification.verifiers.map((verifier) => ({
       ...verifier,
-      lastUsedDaysAgo: getVerifierLastUsedDaysAgo(
-        verifier,
-        verifiersApiResponse,
-      ),
+      lastUsedDaysAgo: getVerifierLastUsedDaysAgo(verifier, verifiersStatuses),
     })),
   }
 }
 
 export function getVerifierLastUsedDaysAgo(
   verifier: OnchainVerifier,
-  verifiersApiResponse: VerifiersApiResponse,
+  verifiersStatuses: VerifierStatuses,
 ): number | undefined {
-  const timestamp = verifiersApiResponse.find(
+  const timestamp = verifiersStatuses.find(
     (entry) => entry.address === verifier.contractAddress.toString(),
   )?.timestamp
