+++ packages/frontend2/src/server/features/zk-catalog/get-verifiers.ts
@@ -1,9 +1,10 @@
 import { getVerifiersFromConfig } from '@l2beat/config/build/src/projects/other/zk-catalog'
-import { VerifiersApiResponse } from '@l2beat/shared-pure'
+import { branded, UnixTime } from '@l2beat/shared-pure'
 import {
   unstable_cache as cache,
   unstable_noStore as noStore,
 } from 'next/cache'
+import { z } from 'zod'
 import { db } from '~/server/database'
 
 export async function getVerifiers() {
@@ -13,7 +14,7 @@ export async function getVerifiers() {
   // issues with custom VOs like UnixTime - that's why we re-parse the data
   // to coerce it back to the correct types
   const cachedVerifiers = await getCachedVerifiersStatus()
-  return VerifiersApiResponse.parse(cachedVerifiers)
+  return VerifierStatuses.parse(cachedVerifiers)
 }
 
 const getCachedVerifiersStatus = cache(
@@ -28,7 +29,7 @@ const getCachedVerifiersStatus = cache(
 
       return {
         address: verifier.contractAddress.toString(),
-        timestamp: status ? status.lastUsed : null,
+        timestamp: status ? status.lastUsed.toNumber() : null,
       }
     })
 
@@ -37,3 +38,13 @@ const getCachedVerifiersStatus = cache(
   ['zk-catalog-verifiers'],
   { revalidate: 60 * 60 },
 )
+
+export const VerifierStatus = z.object({
+  address: z.string(),
+  timestamp: branded(z.number().nullable(), (n) =>
+    n ? new UnixTime(n) : null,
+  ),
+})
+export type VerifierStatus = z.infer<typeof VerifierStatus>
+export const VerifierStatuses = z.array(VerifierStatus)
+export type VerifierStatuses = z.infer<typeof VerifierStatuses>
