+++ packages/shared-pure/src/types/api/index.ts
@@ -9,4 +9,3 @@ export * from './ManuallyVerifiedContrats'
 export * from './TokenTvlApi'
 export * from './TvlApi'
 export * from './VerificationStatus'
-export * from './VerifiersApi'
